requirements.txt
click
simple-youtube-api
pyyaml